added authorization to instructors, items, and musicians routers

+++ routes/instructorRouter.js
@@ -14,7 +14,8 @@ instructorRouter.route('/')
     })
     .catch(err => next(err));
 })
-.post(authenticate.verifyUser, authenticate.verifyAdmin, (req, res, next) => {
+.post(authenticate.verifyUser, (req, res, next) => {
+    req.body.author = req.user._id;
     Instructor.create(req.body)
     .then(instructor => {
         console.log('Instructor Created ', instructor);
@@ -24,11 +25,11 @@ instructorRouter.route('/')
     })
     .catch(err => next(err));
 })
-.put((req, res) => {
+.put(authenticate.verifyUser, (req, res) => {
     res.statusCode = 405;
     res.end('PUT operation not supported on /instructors');
 })
-.delete((req, res, next) => {
+.delete(authenticate.verifyUser, authenticate.verifyAdmin, (req, res, next) => {
     Instructor.deleteMany()
     .then(response => {
         res.statusCode = 200;
@@ -52,23 +53,63 @@ instructorRouter.route('/:instructorId')
     res.statusCode = 405;
     res.end(`POST operation not supported on /instructors/${req.params.instructorId}`);
 })
-.put((req, res, next) => {
-    Instructor.findByIdAndUpdate(req.params.instructorId, {
-        $set: req.body
-    }, { new: true })
+.put(authenticate.verifyUser, (req, res, next) => {
+    console.log(req.user);
+    Instructor.findById(req.params.instructorId)
+    .populate('author')
     .then(instructor => {
-        res.statusCode = 200;
-        res.setHeader('Content-Type', 'application/json');
-        res.json(instructor);
+        if (instructor) {
+            console.log(instructor);
+            console.log(typeof instructor);
+            if (req.user._id.equals(instructor.author._id)) {
+                Instructor.findByIdAndUpdate(req.params.instructorId, {  
+                    $set: req.body
+                }, { new: true })
+                .then(instructor => {
+                    res.statusCode = 200;
+                    res.setHeader('Content-Type', 'application/json'); 
+                    res.json(instructor); 
+                })
+                .catch(err => next(err));
+            } else {
+                const err = new Error('Operation not authorized!');
+                err.status = 403;
+                return next(err);
+            }
+        } else {
+            const err = new Error(`Instructor ${req.params.instructorId} not found`);
+            err.status = 404;
+            return next(err);
+        }
     })
-    .catch(err => next(err));
+    .catch(err => next(err))
 })
-.delete((req, res, next) => {
-   Instructor.findByIdAndDelete(req.params.instructorId)
-    .then(response => {
-        res.statusCode = 200;
-        res.setHeader('Content-Type', 'application/json');
-        res.json(response);
+.delete(authenticate.verifyUser, (req, res, next) => {
+    console.log(req.user);
+    Instructor.findById(req.params.instructorId)
+    .populate('author')
+    .then(instructor => {
+        if (instructor) {
+            console.log(instructor);
+            console.log(typeof instructor);
+            if (req.user._id.equals(instructor.author._id)) {
+                Instructor.findByIdAndDelete(req.params.instructorId)
+                .then(response => {
+                    res.statusCode = 200;
+                    res.setHeader('Content-Type', 'application/json');
+                    res.json(response);
+                })
+                .catch(err => next(err));
+            } else {
+                const err = new Error('Operation not authorized!');
+                err.status = 403;
+                return next(err);
+            }
+        } else {
+            const err = new Error(`Instructor ${req.params.instructorId} not found`);
+            err.status = 404;
+            return next(err);
+        }
     })
     .catch(err => next(err));
 });
