+++ routes/itemRouter.js
@@ -1,5 +1,6 @@
 const express = require('express');
 const Item =  require('../models/items');
+const authenticate = require('../authenticate');
 
 const itemRouter = express.Router();
 
@@ -13,7 +14,8 @@ itemRouter.route('/')
     })
     .catch(err => next(err));
 })
-.post((req, res, next) => {
+.post(authenticate.verifyUser, (req, res, next) => {
+    req.body.author = req.user._id;
     Item.create(req.body)
     .then(item => {
         console.log('Item Created ', item);
@@ -23,11 +25,11 @@ itemRouter.route('/')
     })
     .catch(err => next(err));
 })
-.put((req, res) => {
+.put(authenticate.verifyUser, (req, res) => {
     res.statusCode = 405;
     res.end('PUT operation not supported on /items');
 })
-.delete((req, res, next) => {
+.delete(authenticate.verifyUser, authenticate.verifyAdmin, (req, res, next) => {
     Item.deleteMany()
     .then(response => {
         res.statusCode = 200;
@@ -51,23 +53,63 @@ itemRouter.route('/:itemId')
     res.statusCode = 405;
     res.end(`POST operation not supported on /items/${req.params.itemId}`);
 })
-.put((req, res, next) => {
-    Item.findByIdAndUpdate(req.params.itemId, {
-        $set: req.body
-    }, { new: true })
+.put(authenticate.verifyUser, (req, res, next) => {
+    console.log(req.user);
+    Item.findById(req.params.itemId)
+    .populate('author')
     .then(item => {
-        res.statusCode = 200;
-        res.setHeader('Content-Type', 'application/json');
-        res.json(item);
+        if (item) {
+            console.log(item);
+            console.log(typeof item);
+            if (req.user._id.equals(item.author._id)) {
+                Item.findByIdAndUpdate(req.params.itemId, {  
+                    $set: req.body
+                }, { new: true })
+                .then(item => {
+                    res.statusCode = 200;
+                    res.setHeader('Content-Type', 'application/json'); 
+                    res.json(item); 
+                })
+                .catch(err => next(err));
+            } else {
+                const err = new Error('Operation not authorized!');
+                err.status = 403;
+                return next(err);
+            }
+        } else {
+            const err = new Error(`Item ${req.params.itemId} not found`);
+            err.status = 404;
+            return next(err);
+        }
     })
-    .catch(err => next(err));
+    .catch(err => next(err))
 })
-.delete((req, res, next) => {
-   Item.findByIdAndDelete(req.params.itemId)
-    .then(response => {
-        res.statusCode = 200;
-        res.setHeader('Content-Type', 'application/json');
-        res.json(response);
+.delete(authenticate.verifyUser, (req, res, next) => {
+    console.log(req.user);
+    Item.findById(req.params.itemId)
+    .populate('author')
+    .then(item => {
+        if (item) {
+            console.log(item);
+            console.log(typeof item);
+            if (req.user._id.equals(item.author._id)) {
+                Item.findByIdAndDelete(req.params.itemId)
+                .then(response => {
+                    res.statusCode = 200;
+                    res.setHeader('Content-Type', 'application/json');
+                    res.json(response);
+                })
+                .catch(err => next(err));
+            } else {
+                const err = new Error('Operation not authorized!');
+                err.status = 403;
+                return next(err);
+            }
+        } else {
+            const err = new Error(`Item ${req.params.itemId} not found`);
+            err.status = 404;
+            return next(err);
+        }
     })
     .catch(err => next(err));
 });
